feat: move preference to navbar
* use jotai for topology card size
